+++ src/ontodia/data/sparql/sparqlDataProviderSettings.ts
@@ -6,6 +6,7 @@ export interface SparqlDataProviderSettings {
      * default prefix to be used in every query
      */
     defaultPrefix: string;
+
     /**
      *  property to use as label in schema (classes, properties)
      */
@@ -36,6 +37,11 @@ export interface SparqlDataProviderSettings {
      * query for fetching all information on element: labels, classes, properties
      */
     elementInfoQuery: string;
+
+    /**
+     * Query on all links between said instances. Should return source type target
+     */
+    linksInfoQuery: string;
 
     /**
      * this should return image URL for ?inst as instance and ?linkType for image property IRI
@@ -93,7 +99,39 @@ export interface FullTextSearchSettings {
     extractLabel?: boolean;
 }
 
-export const WikidataSettings: SparqlDataProviderSettings = {
+export const RDFSettings: SparqlDataProviderSettings = {
+    linksInfoQuery: `SELECT ?source ?type ?target
+            WHERE {
+                ?source ?type ?target.
+                VALUES (?source) {\${ids}}
+                VALUES (?target) {\${ids}}                
+            }`,
+
+    defaultPrefix: '',
+
+    schemaLabelProperty: 'rdfs:label',
+    dataLabelProperty: 'rdfs:label',
+
+    fullTextSearch: {
+        prefix: '',
+        queryPattern: ``,
+    },
+
+    classTreeQuery: ``,
+
+    linkTypesPattern: ``,
+
+    elementInfoQuery: ``,
+    imageQueryPattern: ``,
+
+    linkTypesOfQuery: ``,
+    filterRefElementLinkPattern: '',
+    filterTypePattern: ``,
+    filterAdditionalRestriction: ``,
+    filterElementInfoPattern: ``,
+};
+
+const WikidataSettingsOverride: Partial<SparqlDataProviderSettings> = {
     defaultPrefix:
         `PREFIX rdfs: <http://www.w3.org/2000/01/rdf-schema#>
  PREFIX rdf:  <http://www.w3.org/1999/02/22-rdf-syntax-ns#>
@@ -185,9 +223,12 @@ export const WikidataSettings: SparqlDataProviderSettings = {
                 BIND (coalesce(?foundClass, owl:Thing) as ?class)
                 OPTIONAL {?inst rdfs:label ?label}
 `,
+
 };
 
-export const OWLRDFSSettings: SparqlDataProviderSettings = {
+export const WikidataSettings: SparqlDataProviderSettings = {...RDFSettings, ...WikidataSettingsOverride};
+
+export const OWLRDFSSettingsOverride: Partial<SparqlDataProviderSettings> = {
     defaultPrefix:
         `PREFIX rdfs: <http://www.w3.org/2000/01/rdf-schema#>
  PREFIX rdf:  <http://www.w3.org/1999/02/22-rdf-syntax-ns#>
@@ -249,6 +290,8 @@ export const OWLRDFSSettings: SparqlDataProviderSettings = {
                 OPTIONAL {?inst \${dataLabelProperty} ?label}`,
     filterAdditionalRestriction: '',
 };
+
+export const OWLRDFSSettings: SparqlDataProviderSettings = {...RDFSettings, ...OWLRDFSSettingsOverride};
 
 const OWLStatsOverride: Partial<SparqlDataProviderSettings> = {
     classTreeQuery: `