+++ src/ontodia/diagram/linkView.ts
@@ -5,6 +5,7 @@ import { isIE11 } from '../viewUtils/detectBrowser';
 
 import { Link, FatLinkType } from './elements';
 import { DiagramView } from './view';
+import { LinkStyle } from '../customization/props';
 
 export class LinkView extends joint.dia.LinkView {
     model: Link;
@@ -19,6 +20,7 @@ export class LinkView extends joint.dia.LinkView {
         this.listenTo(this.model, 'change:layoutOnly', this.updateLabel);
         this.listenTo(this.model, 'updateRouting', this.onRoutingUpdate);
     }
+
     render(): LinkView {
         if (!this.view && this.paper && this.paper.diagramView) {
             this.setView(this.paper.diagramView);
@@ -26,6 +28,7 @@ export class LinkView extends joint.dia.LinkView {
         const result: any = super.render();
         return result;
     }
+
     getTypeModel(): FatLinkType {
         return this.view.model.getLinkType(this.model.get('typeId'));
     }
@@ -58,32 +61,54 @@ export class LinkView extends joint.dia.LinkView {
     }
 
     private updateLabelWithOptions(options?: { silent?: boolean }) {
-        const linkTypeId: string = this.model.get('typeId');
-        const typeModel = this.view.model.getLinkType(linkTypeId);
-
-        const style = this.view.getLinkStyle(this.model.get('typeId'));
+        const template = this.view.getLinkTemplate(this.model.typeId);
+        const style = template.renderLink(this.model.template);
         merge(style, {connection: {'stroke-dasharray': this.model.layoutOnly ? '5,5' : null}});
 
-        let linkAttributes: joint.dia.LinkAttributes = {
-            labels: style.labels,
+        const linkAttributes: joint.dia.LinkAttributes = {
+            labels: this.createLabelsAttributes(style),
             connector: style.connector,
-            router: style.router,
+            router: template.router,
             z: 0,
         };
         if (style.connection) {
             merge(linkAttributes, {attrs: {'.connection': style.connection}});
         }
 
-        const showLabels = typeModel && typeModel.get('showLabel');
-        const labelAttributes = showLabels ? [{
-            position: 0.5,
-            attrs: {text: {
-                text: this.view.getLinkLabel(linkTypeId).text,
-            }},
-        }] : [];
-
-        merge(linkAttributes, {labels: labelAttributes});
         this.model.set(linkAttributes, options);
+    }
+
+    private createLabelsAttributes = (style: LinkStyle): joint.dia.LinkLabelAttributes[] => {
+        const linkTypeId = this.model.typeId;
+        const typeModel = this.view.model.getLinkType(linkTypeId);
+        const showLabels = typeModel && typeModel.get('showLabel');
+        const labelAttributes: joint.dia.LinkLabelAttributes[] = [];
+
+        if (showLabels) {
+            if (style.label) {
+                const label = {};
+                const value = style.label.attrs && style.label.attrs.text ? style.label.attrs.text.text : undefined;
+                const text = value ? this.view.getLocalizedText(value).text : this.view.getLinkLabel(linkTypeId).text;
+                merge(label, style.label, {position: style.label.position || 0.5, attrs: {text: {text}}});
+                labelAttributes.push(label);
+            } else {
+                const text = this.view.getLinkLabel(linkTypeId).text;
+                const label = {position: 0.5, attrs: {text: {text}}};
+                labelAttributes.push(label);
+            }
+
+            if (style.properties) {
+                style.properties.forEach(prop => {
+                    const property = {};
+                    const value = prop.attrs && prop.attrs.text ? prop.attrs.text.text : undefined;
+                    const text = value ? this.view.getLocalizedText(value).text : '';
+                    merge(property, prop, {position: prop.position || 0.5, attrs: {text: {text}}});
+                    labelAttributes.push(property);
+                });
+            }
+        }
+
+        return labelAttributes;
     }
 }